+++ app/app.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+// TODO: Set up useGqlError() to handle GQL errors globally
 const { GQL_HOST: gqlHost, channelToken } = useRuntimeConfig().public;
 const { locale } = useI18n();
 const orderStore = useOrderStore();